requirements.txt
# Production requirements for Railway deployment
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-cors-headers>=4.0.0
dj-database-url>=2.0.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
whitenoise>=6.0.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
django-ses>=4.0.0
django-filter>=23.0.0

# AWS Services
boto3>=1.26.0
botocore>=1.29.0

# Data Processing & Analysis
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0

# Document Processing
pymupdf>=1.23.0
docx2txt>=0.8
python-docx>=0.8.11
reportlab>=4.0.0

# AI & ML
openai>=1.0.0
opencv-python>=4.8.0

# Communication
websockets>=11.0.0
requests>=2.28.0
twilio>=8.0.0
sendgrid>=6.0.0

# Payment Processing
razorpay>=1.4.0
setuptools>=68.0.0
wheel>=0.38.0
packaging>=21.0.0

# Task Processing
celery>=5.3.0
redis>=5.0.0

# Additional utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Healthcare Specific
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
